requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
flatbuffers==25.12.19
fonttools==4.53.1
fsspec==2024.2.0
grpcio==1.65.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mediapipe==0.10.14
mkl==2021.4.0
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.13.0.92
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.23
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sounddevice==0.5.5
sympy==1.12
tbb==2021.11.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.59
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
